fix(android): read hwcap using getauxval (v1)
There are Android devices where reading /proc/self/auxv fails
with permission denied. Android is increasingly preventing apps
and tools from accessing system functionality.

Go uses /proc/self/auxv to fill cpu.ARM64. In turn, it uses
the values in this data structure to choose whether to enable
an assembly implementation of AES as well as to choose the
order of the cipher suites in the TLS ClientHello.

Unfortunately, as documented in https://github.com/ooni/probe/issues/1444,
reordering the ciphers in the ClientHello triggers censorship
in Iran. Because new(?) Android devices do not allow Go to access
/proc/self/auxv, we therefore see that OONI is not WAI.

The most obvious fix to this issue could seem to teach crypto/tls to
always put AES ciphers on top. Unfortunately, the pure Go AES
implementation in the standard library is not constant time (while
the one in assembly is constant time). Therefore, forcing AES to
always be at the top of the ciphers list does not seem right.

Thus, we can say that our real core problem is that Go is not able
to know whether we support AES and PMULL on android/arm64.

This diff attempts to solve this problem. The solution I have
chosen is to implement a cpuproxy package exposing predicate
functions describing the CPU's support for AES and PMULL.

This cpuproxy package will use the existing logic for every GOOS
and GOARCH except android/arm64. When we are in android/arm64,
we use `getauxval(3)` to obtain the hardware capabilities and we
use this information to implement the predicates.

My initial attempt at solving this problem was to call `getauxval(3)`
directly during the initialization of the `runtime` package. But it
seems that when you use CGO you depend on the `syscall` package, and
in turn `syscall` depends on `runtime`. So, when I was compiling OONI
for Android, I always got a circular-import hard failure.

Another approach could be that of reading `/proc/cpuinfo` inside of
the `runtime` initialization. I did not explore it.

This solution based on `libc` and `getauxval(3)` is not perfect but we
already build OONI for Android with CGO enabled, so we should be fine
in terms of the functionality that we require as OONI.

I also adjusted the tests to reflect the fact that now there are more
packages that depend on CGO (when using Android).