fix(s/g/b/deps_test.go): make sure tests are still passing.
I extracted and adapted this diff from 62177de2b42c35bee645909cc074a2b6c4b67d5a
with the intent of coming up with a more manageable patch set.

The idea here is to fix the tests because we have changed the imports
expected by some packages inside of src/crypto to use i/cpuproxy.

The original issue describing this problem in its full extent was
https://github.com/ooni/probe/issues/1444.

We're now working to forward-port this patch to go1.17.

The current issue describing this problem as of go1.17 is
https://github.com/ooni/probe/issues/1863.

Is there a better way?